backend/requirements.txt
# ============================================================================
# ChatPop Backend - Frozen Requirements
# ============================================================================
# This file contains all dependencies with exact version numbers frozen to
# ensure reproducible builds. Generated from pip freeze.
# ============================================================================

# Django and REST Framework
Django==5.0.14
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-constance==4.3.2
django-redis==6.0.0

# ASGI Server
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0

# Database
psycopg2-binary==2.9.10
pgvector==0.4.1

# Redis
redis==5.0.8
hiredis==2.2.3

# Environment
python-dotenv==1.0.1

# Payments
stripe==7.14.0

# Utilities
PyJWT==2.10.1

# Media Processing
ffmpeg-python==0.2.0  # Voice message transcoding for iOS compatibility
pillow==12.0.0  # Image processing for photo analysis

# Photo Analysis & AI
openai==2.5.0  # Photo analysis for chat generation
ImageHash==4.3.2  # Perceptual image hashing for duplicate detection
numpy==2.3.4  # Required by ImageHash
scipy==1.16.2  # Required by ImageHash
PyWavelets==1.9.0  # Required by ImageHash

# API Documentation
drf-spectacular==0.28.0

# Testing
pytest==8.4.2
pytest-django==4.11.1
pytest-cov==6.0.0
allure-pytest==2.15.0

# HTTP Clients
requests==2.32.5
httpx==0.28.1

# ============================================================================
# Auto-installed Dependencies (transitive dependencies)
# ============================================================================

# Core dependencies
asgiref==3.9.2
autobahn==24.4.2
certifi==2025.8.3
pydantic==2.12.3
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
Twisted==25.5.0

# Testing dependencies
allure-python-commons==2.15.0
attrs==25.3.0
coverage==7.11.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0

# Cryptography and security
cryptography==46.0.2
cffi==2.0.0
pycparser==2.23
pyOpenSSL==25.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
service-identity==24.2.0

# HTTP and networking
anyio==4.11.0
h11==0.16.0
httpcore==1.0.9
idna==3.10
sniffio==1.3.1
urllib3==2.5.0
hyperlink==21.0.0
websocket==0.2.1

# Twisted dependencies
Automat==25.4.16
constantly==23.10.4
incremental==24.7.2
txaio==25.9.2
zope.event==6.0
zope.interface==8.0.1

# Async and concurrency
gevent==25.9.1
greenlet==3.2.4

# Pydantic dependencies
annotated-types==0.7.0
pydantic_core==2.41.4
typing-inspection==0.4.2
typing_extensions==4.15.0

# OpenAI dependencies
distro==1.9.0
jiter==0.11.1
tqdm==4.67.1

# drf-spectacular dependencies
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.36.2
rpds-py==0.27.1
uritemplate==4.2.0

# Other utilities
charset-normalizer==3.4.3
future==1.0.0
msgpack==1.1.1
Pygments==2.19.2
setuptools==80.9.0
pyacrcloud
